+++ config.cmake
@@ -13,10 +13,13 @@ option(WITH_MPI "Whether DFTB+ should support MPI-parallelism" FALSE)
 # If you build an MPI-parallised binary, consider to set WITH_OMP (OpenMP thread parallelisaton) to
 # FALSE unless you want hybrid parallelisation (for experts only).
 
-option(WITH_ELSI "Whether DFTB+ with MPI-parallelism should use the ELSI libraries" FALSE)
-# Works only with MPI-parallel build.
+option(WITH_GPU "Whether DFTB+ should support GPU-acceleration" FALSE)
+# For serial builds, the GPU support requires the MAGMA library. For MPI parallel builds it
+# requires the ELSI library built with GPU support.
 
-option(WITH_GPU "Whether DFTB+ should support GPU-acceleration via the MAGMA-library" FALSE)
+option(WITH_ELSI "Whether DFTB+ with MPI-parallelism should use the ELSI libraries" FALSE)
+# Works only with MPI-parallel build. If WITH_GPU was selected above, the ELSI library must be
+# enabled (and must have been built with GPU support).
 
 option(WITH_TRANSPORT "Whether transport via libNEGF should be included." FALSE)
 # Works only when building static libraries (see option BUILD_SHARED_LIBS)