+++ src/dftbp/api/mm/mmapi.F90
@@ -12,11 +12,11 @@ module dftbp_mmapi
   use iso_fortran_env, only : output_unit
   use dftbp_common_accuracy, only : dp
   use dftbp_common_environment, only : TEnvironment, TEnvironment_init
-  use dftbp_common_globalenv, only : initGlobalEnv, destructGlobalEnv, instanceSafeBuild
+  use dftbp_common_globalenv, only : initGlobalEnv, destructGlobalEnv, instanceSafeBuild, withMpi
   use dftbp_dftbplus_hsdhelpers, only : doPostParseJobs
   use dftbp_dftbplus_initprogram, only: TDftbPlusMain
   use dftbp_dftbplus_inputdata, only : TInputData
-  use dftbp_dftbplus_mainapi, only : doOneTdStep, checkSpeciesNames, nrOfAtoms,&
+  use dftbp_dftbplus_mainapi, only : doOneTdStep, checkSpeciesNames, nrOfAtoms, nrOfKPoints,&
       & setExternalPotential, getTdForces, setTdCoordsAndVelos, setTdElectricField,&
       & initializeTimeProp, updateDataDependentOnSpeciesOrdering, getAtomicMasses,&
       & getGrossCharges, getElStatPotential, getExtChargeGradients, getStressTensor, getGradients,&
@@ -102,6 +102,8 @@ module dftbp_mmapi
     procedure :: getElStatPotential => TDftbPlus_getElStatPotential
     !> Return the number of DFTB+ atoms in the system
     procedure :: nrOfAtoms => TDftbPlus_nrOfAtoms
+    !> Return the number of k-points in the DFTB+ calculation (1 if non-repeating)
+    procedure :: nrOfKPoints => TDftbPlus_nrOfKPoints
     !> Check that the list of species names has not changed
     procedure :: checkSpeciesNames => TDftbPlus_checkSpeciesNames
     !> Replace species and redefine all quantities that depend on it
@@ -120,6 +122,8 @@ module dftbp_mmapi
     procedure, private :: checkInit => TDftbPlus_checkInit
     !> Return the masses for each atom in the system
     procedure :: getAtomicMasses => TDftbPlus_getAtomicMasses
+    !> Return the number of basis functions for each atom in the system
+    procedure :: getNOrbitalsOnAtoms => TDftbPlus_getNOrbAtoms
   end type TDftbPlus
 
 
@@ -284,6 +288,10 @@ contains
       end if
       nInstance_ = 1
     #:endif
+
+    if (present(mpiComm) .and. .not. withMpi) then
+      call error("MPI Communicator supplied to initialise serial DFTB+ instance")
+    end if
 
     if (present(outputUnit)) then
       stdOut = outputUnit
@@ -531,6 +539,7 @@ contains
 
   end subroutine TDftbPlus_getStressTensor
 
+
   !> Returns the gradients on the external charges.
   !>
   !> This function may only be called if TDftbPlus_setExternalCharges was called before it
@@ -601,6 +610,22 @@ contains
   end function TDftbPlus_nrOfAtoms
 
 
+  !> Returns the nr. of k-points describing the system.
+  function TDftbPlus_nrOfKPoints(this) result(nKpts)
+
+    !> Instance
+    class(TDftbPlus), intent(in) :: this
+
+    !> Nr. of k-points
+    integer :: nKpts
+
+    call this%checkInit()
+
+    nKpts = nrOfKPoints(this%main)
+
+  end function TDftbPlus_nrOfKPoints
+
+
   !> Returns the atomic masses for each atom in the system.
   subroutine TDftbPlus_getAtomicMasses(this, mass)
 
@@ -615,6 +640,20 @@ contains
     call getAtomicMasses(this%main, mass)
 
   end subroutine TDftbPlus_getAtomicMasses
+
+
+  !> Returns the number of orbitals for each atom in the system
+  subroutine TDftbPlus_getNOrbAtoms(this, nOrbs)
+
+    !> Instance
+    class(TDftbPlus), intent(inout) :: this
+
+    !> Number of basis functions associated with each atom
+    integer, intent(out) :: nOrbs(:)
+
+    nOrbs(:) = this%main%orb%nOrbAtom
+
+  end subroutine TDftbPlus_getNOrbAtoms
 
 
   !> Checks whether the type is already initialized and stops the code if not.
@@ -703,6 +742,7 @@ contains
     end do
 
   end subroutine convertAtomTypesToSpecies
+
 
   !> Check whether speciesNames has changed between calls to DFTB+
   subroutine TDftbPlus_checkSpeciesNames(this, inputSpeciesNames)
@@ -802,7 +842,7 @@ contains
     !> Instance
     class(TDftbPlus), intent(inout) :: this
 
-    ! electric field components
+    !> electric field components
     real(dp), intent(in) :: field(3)
 
     if (allocated(this%main%solvation)) then
@@ -816,15 +856,16 @@ contains
   end subroutine TDftbPlus_setTdElectricField
 
 
+  !> Set atomic coordinates and velocities for MD
   subroutine TDftbPlus_setTdCoordsAndVelos(this, coords, velos)
 
     !> Instance
     class(TDftbPlus), intent(inout) :: this
 
-    ! coordinates
+    !> coordinates
     real(dp), intent(in) :: coords(3, this%main%nAtom)
 
-    ! velocities
+    !> velocities
     real(dp), intent(in) :: velos(3, this%main%nAtom)
 
     call setTdCoordsAndVelos(this%main, coords, velos)
@@ -832,6 +873,7 @@ contains
   end subroutine TDftbPlus_setTdCoordsAndVelos
 
 
+  !> Returns forces from time dependent propagation
   subroutine TDftbPlus_getTdForces(this, forces)
 
     !> Instance
@@ -843,6 +885,5 @@ contains
     call getTdForces(this%main, forces)
 
   end subroutine TDftbPlus_getTdForces
-
 
 end module dftbp_mmapi