+++ src/dftbp/dftbplus/mainapi.F90
@@ -31,7 +31,7 @@ module dftbp_dftbplus_mainapi
   private
   public :: setGeometry, setQDepExtPotProxy, setExternalPotential, setExternalCharges
   public :: getEnergy, getGradients, getExtChargeGradients, getGrossCharges, getElStatPotential
-  public :: getStressTensor, nrOfAtoms, getAtomicMasses
+  public :: getStressTensor, nrOfAtoms, nrOfKPoints, getAtomicMasses
   public :: updateDataDependentOnSpeciesOrdering, checkSpeciesNames
   public :: initializeTimeProp, doOneTdStep, setTdElectricField, setTdCoordsAndVelos, getTdForces
 
@@ -323,6 +323,19 @@ contains
   end function nrOfAtoms
 
 
+  !> Obtains number of k-points in the system (1 if not a repeating structure)
+  function nrOfKPoints(main)
+
+    !> Instance
+    type(TDftbPlusMain), intent(in) :: main
+
+    integer :: nrOfKPoints
+
+    nrOfKPoints = main%nKPoint
+
+  end function nrOfKPoints
+
+
   !> Check that the order of speciesName remains constant Keeping speciesNames constant avoids the
   !> need to reset all of atomEigVal, referenceN0, speciesMass and SK parameters
   !>
@@ -587,7 +600,7 @@ contains
   end subroutine setTdCoordsAndVelos
 
 
-  !> gets atomic forces from td propagation
+  !> gets atomic forces from time dependent propagation
   subroutine getTdForces(main, forces)
 
     !> Instance
@@ -597,6 +610,7 @@ contains
     real(dp), intent(out) :: forces(:,:)
 
     forces(:,:) = main%electronDynamics%totalForce
+
   end subroutine getTdForces